room取得のタイミングを変更
app.jsのcomponentDidMountで取りに行っているが、roomsコンポーネントで取りに行くように変更
room一覧からページが移ると、chat room channelが閉じるように変更
